chore: rename the login.tsx file to meet standards
